fix: Re-enable fast reload
