+++ src/api/policy.ts
@@ -1,4 +1,5 @@
 import { BASE_URL } from '@/constants';
+import { PolicyState } from '@/reducers/policyReducer';
 
 export async function fetchPolicyById(country: string, policyId: string) {
   const url = `${BASE_URL}/${country}/policy/${policyId}`;
@@ -35,4 +36,28 @@ export async function createPolicy(data: any) {
   }
 
   return res.json();
+}
+
+interface PolicyCreationPayload {
+  label?: string;
+  data: Record<string, any>;
+}
+
+export function serializePolicyCreationPayload(policy: PolicyState): PolicyCreationPayload {
+  const { label, params } = policy;
+
+  // Fill payload with keys we already know
+  const payload = {
+    label,
+    data: {} as Record<string, any>,
+  };
+
+  // Convert params and their valueIntervals into expected JSON format
+  params.forEach((param) => {
+    payload.data[param.name] = param.values.reduce((acc, cur) => {
+      return { ...acc, [`${cur.startDate}..${cur.endDate}`]: cur.value };
+    }, {});
+  });
+
+  return payload;
 }
