+++ src/reducers/policyReducer.ts
@@ -25,30 +25,6 @@ const initialState: PolicyState = {
   params: [],
 };
 
-interface PolicyCreationPayload {
-  label?: string;
-  data: Record<string, any>;
-}
-
-export function serializePolicyCreationPayload(policy: PolicyState): PolicyCreationPayload {
-  const { label, params } = policy;
-
-  // Fill payload with keys we already know
-  const payload = {
-    label,
-    data: {} as Record<string, any>,
-  };
-
-  // Convert params and their valueIntervals into expected JSON format
-  params.forEach((param) => {
-    payload.data[param.name] = param.values.reduce((acc, cur) => {
-      return { ...acc, [`${cur.startDate}..${cur.endDate}`]: cur.value };
-    }, {});
-  });
-
-  return payload;
-}
-
 // TODO: Some of this code is a bit naive
 export const policySlice = createSlice({
   name: 'policy',
